+++ airgeddon.sh
@@ -2858,7 +2858,7 @@ function handshake_capture_check() {
 	local time_counter=0
 	while true; do
 		sleep 5
-		if check_bssid_in_captured_file "${tmpdir}${standardhandshake_filename}" "silent"; then
+		if check_bssid_in_captured_file "${tmpdir}${standardhandshake_filename}" "silent" "only_handshake"; then
 			break
 		fi
 
@@ -6947,14 +6947,23 @@ function check_valid_file_to_clean() {
 	return 0
 }
 
-#Check if a bssid is present on a capture file to know if there is a Handshake with that bssid
+#Check if a bssid is present on a capture file to know if there is a Handshake/PMKID with that bssid
 function check_bssid_in_captured_file() {
 
 	debug_print
 
+	local nets_from_file
 	nets_from_file=$(echo "1" | aircrack-ng "${1}" 2> /dev/null | grep -E "WPA \([1-9][0-9]? handshake" | awk '{ saved = $1; $1 = ""; print substr($0, 2) }')
 
-	if [ "${2}" != "silent" ]; then
+	if [ "${3}" = "also_pmkid" ]; then
+		get_aircrack_version
+		if compare_floats_greater_or_equal "${aircrack_version}" "${aircrack_pmkid_version}"; then
+			local nets_from_file2
+			nets_from_file2=$(echo "1" | aircrack-ng "${1}" 2> /dev/null | grep -E "handshake, with PMKID" | awk '{ saved = $1; $1 = ""; print substr($0, 2) }')
+		fi
+	fi
+
+	if [[ "${2}" = "showing_msgs" ]] && [[ "${3}" = "only_handshake" ]]; then
 		echo
 		if [ "${nets_from_file}" = "" ]; then
 			if [ ! -f "${1}" ]; then
@@ -6969,6 +6978,9 @@ function check_bssid_in_captured_file() {
 	fi
 
 	declare -A bssids_detected
+	declare -A bssids_detected_pmkid
+
+	local option_counter
 	option_counter=0
 	for item in ${nets_from_file}; do
 		if [[ ${item} =~ ^[0-9a-fA-F]{2}: ]]; then
@@ -6977,20 +6989,55 @@ function check_bssid_in_captured_file() {
 		fi
 	done
 
+	if [[ "${3}" = "also_pmkid" ]] && [[ -n "${nets_from_file2}" ]]; then
+		option_counter=0
+		for item in ${nets_from_file2}; do
+			if [[ ${item} =~ ^[0-9a-fA-F]{2}: ]]; then
+				option_counter=$((option_counter + 1))
+				bssids_detected_pmkid[${option_counter}]=${item}
+			fi
+		done
+	fi
+
+	local handshake_captured=0
+	local pmkid_captured=0
+
 	for targetbssid in "${bssids_detected[@]}"; do
 		if [ "${bssid}" = "${targetbssid}" ]; then
-			if [ "${2}" != "silent" ]; then
-				language_strings "${language}" 322 "yellow"
-			fi
-			return 0
+			handshake_captured=1
+			break
 		fi
 	done
 
-	if [ "${2}" != "silent" ]; then
-		language_strings "${language}" 323 "red"
-		language_strings "${language}" 115 "read"
+	if [[ "${3}" = "also_pmkid" ]] && [[ -n "${nets_from_file2}" ]]; then
+		for targetbssid in "${bssids_detected_pmkid[@]}"; do
+			if [ "${bssid}" = "${targetbssid}" ]; then
+				pmkid_captured=1
+				break
+			fi
+		done
 	fi
-	return 1
+
+	if [[ "${handshake_captured}" = "1" ]] && [[ "${pmkid_captured}" = "0" ]]; then
+		if [[ "${2}" = "showing_msgs" ]] && [[ "${3}" = "only_handshake" ]]; then
+			language_strings "${language}" 322 "yellow"
+		fi
+		return 0
+	elif [[ "${handshake_captured}" = "0" ]] && [[ "${pmkid_captured}" = "1" ]]; then
+		echo
+		language_strings "${language}" 680 "yellow"
+		return 0
+	elif [[ "${handshake_captured}" = "1" ]] && [[ "${pmkid_captured}" = "1" ]]; then
+		echo
+		language_strings "${language}" 681 "yellow"
+		return 0
+	else
+		if [[ "${2}" = "showing_msgs" ]] && [[ "${3}" = "only_handshake" ]]; then
+			language_strings "${language}" 323 "red"
+			language_strings "${language}" 115 "read"
+		fi
+		return 1
+	fi
 }
 
 #Set the target vars to a bssid selecting them from a capture file which has a Handshake/PMKID
@@ -7006,6 +7053,7 @@ function select_wpa_bssid_target_from_captured_file() {
 		language_strings "${language}" 115 "read"
 	fi
 
+	local nets_from_file
 	if [ "${2}" = "only_handshake" ]; then
 		nets_from_file=$(echo "1" | aircrack-ng "${1}" 2> /dev/null | grep -E "WPA \([1-9][0-9]? handshake" | awk '{ saved = $1; $1 = ""; print substr($0, 2) }')
 	else
@@ -11318,7 +11366,7 @@ function capture_handshake_evil_twin() {
 
 	handshake_capture_check
 
-	if check_bssid_in_captured_file "${tmpdir}${standardhandshake_filename}" "silent"; then
+	if check_bssid_in_captured_file "${tmpdir}${standardhandshake_filename}" "silent" "only_handshake"; then
 
 		handshakepath="${default_save_path}"
 		handshakefilename="handshake-${bssid}.cap"
@@ -11862,12 +11910,14 @@ function launch_handshake_capture() {
 	fi
 
 	handshake_capture_check
-	if check_bssid_in_captured_file "${tmpdir}${standardhandshake_filename}" "silent"; then
+
+	if check_bssid_in_captured_file "${tmpdir}${standardhandshake_filename}" "showing_msgs" "also_pmkid"; then
 
 		handshakepath="${default_save_path}"
 		handshakefilename="handshake-${bssid}.cap"
 		handshakepath="${handshakepath}${handshakefilename}"
 
+		echo
 		language_strings "${language}" 162 "yellow"
 		validpath=1
 		while [[ "${validpath}" != "0" ]]; do
@@ -11942,6 +11992,7 @@ function launch_pmkid_capture() {
 		pmkidfilename="pmkid-${bssid}.txt"
 		pmkidpath="${pmkidpath}${pmkidfilename}"
 
+		echo
 		language_strings "${language}" 162 "yellow"
 		validpath=1
 		while [[ "${validpath}" != "0" ]]; do
@@ -12570,7 +12621,7 @@ function et_prerequisites() {
 		retry_handshake_capture=0
 		retrying_handshake_capture=0
 
-		if ! check_bssid_in_captured_file "${et_handshake}"; then
+		if ! check_bssid_in_captured_file "${et_handshake}" "showing_msgs" "only_handshake"; then
 			return_to_et_main_menu=1
 			return
 		fi